requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
pydantic
openai
vllm